+++ CHANGELOG.md
@@ -231,7 +231,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Map data json [#12](https://github.com/policy-design-lab/pdl-frontend/issues/12)
 - Final landing page changes for initial milestone [#15](https://github.com/policy-design-lab/pdl-frontend/issues/15)
 
-[0.18.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.16.0...0.18.0
+[0.18.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.17.0...0.18.0
 [0.17.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.16.0...0.17.0
 [0.16.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.15.0...0.16.0
 [0.15.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.14.0...0.15.0